requirements.txt
# Core Azure SDKs
azure-identity==1.19.0
azure-search-documents==11.6.0b7
azure-core>=1.31.0

# LangChain ecosystem
langchain==0.3.7
langchain-community==0.3.7
langchain-core>=0.3.17,<0.4.0

# OpenAI
openai>=1.40.0

# Document processing
pypdf==4.3.1
python-docx==1.1.2
python-pptx==1.0.2
markdown==3.7
beautifulsoup4==4.12.3
lxml==5.3.0

# Text processing
tiktoken==0.8.0
nltk==3.9.1
sentence-transformers==3.2.1

# Web application
streamlit==1.40.1
streamlit-chat==0.1.1

# Data handling
pandas==2.2.3
numpy>=1.26.0,<2.0.0

# Environment & Configuration
python-dotenv==1.0.1
pyyaml==6.0.2
pydantic==2.10.2
pydantic-settings==2.6.1

# Utilities
tenacity==9.0.0
requests==2.32.3
aiohttp==3.11.7

# Logging & Monitoring
loguru==0.7.2
python-json-logger==3.1.0

# Testing
pytest==8.3.3
pytest-cov==6.0.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Code quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Jupyter (for notebooks/demos)
jupyter==1.1.1
ipykernel==6.29.5
